completed till rm_inventory

+++ src/routes/incoming_steel/incoming_steel_model.rs
@@ -12,7 +12,7 @@ pub struct IncomingSteel {
     pub heat_no: String,
     pub heat_code: Option<String>,
     pub jominy_value: Option<String>,
-    pub received_qty: i64,
+    pub received_qty: f64,
     pub heat_status: Option<String>,
     pub created_by: Option<String>,
     pub created_on: DateTime<Local>,
